Rename linuxdeploy.cpp and linuxdeploy.h to core.cpp and core.h
