+++ src/core.h
